Implement serialization by CRTP to allow consumption of binding-specific deployment parameters in a predefined way.
- InputStream & OutputStream are templates now that require providion of
the "real" input or output stream at instantiation time.

- Struct implements structs by using std::tuple. This allows to
serialize structs in a generic way (similar but not identical to what is
done for variants).

- Separate TypeOutputStream from OutputStream as there may/will be
bindings which do not need it (e.g. SOME/IP). Implement TypeOutputStream
in the same way as OutputStream (CRTP).

- Summarize all implementation details of variants in Variant.h instead
of distributing them over SerializableVariant.h and
SerializableVariant.hpp.

- Create base classes Deployable and Deployment for handling
deployments. A Deployment follows the structure of structs and variants
(basically using std::tuple to hold all members) in order to allow to
distribute the deployment in the same way as the value information.

Signed-off-by: Lutz Bichler <Lutz.Bichler@bmw.de>