GLIPCI-453: Fixed bug in CommonAPI::Runtime::LoadLibrary function w/ paths in .ini
The bug caused a faulty addition of ".so" to a path pointing to a shared
gluecode-library specified in the .ini file, if the path contained the
term ".so" in between and not only at the end.

Example:
[proxy]
local:[...]:[...]=[...]/org.genivi.commonapi.someip.verification/build/libSomeIPGlue.so

In this case the additional .so would have been added as
"commonapi.somip" contains the term ".so", which is wrong.

Change-Id: I2a3b0d15c41a86a0432d828dfa46e1f51ef646b6